% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_cranpkg}
\alias{read_cranpkg}
\title{Fetch the source of a pkg into a tibble}
\usage{
read_cranpkg(pkg, verbose = FALSE)
}
\arguments{
\item{pkg}{string, name of a package.}

\item{verbose}{boolean, display debug info.}
}
\value{
a tibble with the source code of all R files, DESCRIPTION and NAMESPACE
}
\description{
This function fetches the source code of a package and dump them into a tibble.
}
