% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_tarball}
\alias{read_tarball}
\title{Dump the source from a package tarball into a tibble}
\usage{
read_tarball(tarball_path, verbose = FALSE, clean_up = TRUE)
}
\arguments{
\item{tarball_path}{string, path to a tarball in .tar.gz.}

\item{verbose}{boolean, display debug info.}

\item{clean_up}{boolean, remove extracted files.}
}
\value{
a tibble with the source code of all R files, DESCRIPTION and NAMESPACE
}
\description{
This function dump the content of a tarball containing an R package into a tibble.
}
