% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_tarball_meta}
\alias{read_tarball_meta}
\title{Extract package name from tarball}
\usage{
read_tarball_meta(tarball_path)
}
\arguments{
\item{tarball_path}{string, path to a tarball in .tar.gz.}

\item{verbose}{boolean, display debug info.}
}
\value{
package name
}
\description{
This function returns the package name from a tarball containing an R package.
}
