% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_tarball_meta}
\alias{read_tarball_meta}
\title{Extract fields from tarball}
\usage{
read_tarball_meta(tarball_path, field = c("Package", "Bundle"))
}
\arguments{
\item{tarball_path}{string, path to a tarball in .tar.gz.}

\item{field}{vector of fields to extract. By default, this function extracts package name.}
}
\value{
field values
}
\description{
This function returns the package name from a tarball containing an R package.
}
